chore(rename): add EarnHistory to model and selectors
