+++ packages/blockchain-wallet-v4-frontend/src/scenes/EarnHistory/EarnHistory.model.tsx
