+++ packages/blockchain-wallet-v4-frontend/src/scenes/EarnHistory/index.tsx
@@ -13,7 +13,7 @@ import { useRemote } from 'hooks'
 import EarnHeader from '../Earn/Earn.template.header'
 import CoinFilter from './CoinFilter'
 import DownloadTransactions from './DownloadTransactions'
-import { getData } from './selectors'
+import { getData } from './EarnHistory.selectors'
 import Tabs from './Tabs'
 import Loading from './template.loading'
 import TransactionList from './template.success'
